Move ctcDeploy() from integration-tests/ to creditcoin-js/
so we can deduplicate its usage later.

Bump creditcoin-js to v0.10.0
