+++ creditcoin-js/src/examples/ctc/contracts/SafeMath.sol/SafeMath.json
