+++ creditcoin-js/src/examples/ctc/typechain/CreditcoinBase.d.ts
