+++ creditcoin-js/src/examples/ctc/typechain/Erc20.d.ts
