+++ creditcoin-js/src/examples/ctc/typechain/Erc20Plus.d.ts
