+++ creditcoin-js/src/examples/ctc/typechain/Owned.d.ts
