+++ creditcoin-js/src/examples/ctc/typechain/common.d.ts
