+++ creditcoin-js/src/examples/ctc/typechain/factories/CreditcoinBase__factory.ts
