+++ creditcoin-js/src/examples/ctc/typechain/factories/Erc20Plus__factory.ts
