+++ creditcoin-js/src/examples/ctc/typechain/factories/Erc20__factory.ts
