+++ creditcoin-js/src/examples/ctc/typechain/factories/Owned__factory.ts
