+++ creditcoin-js/src/examples/ctc/typechain/hardhat.d.ts
