+++ scripts/cc-cli/package.json
@@ -27,7 +27,7 @@
   "dependencies": {
     "cli-table3": "^0.6.3",
     "commander": "^11.0.0",
-    "creditcoin-js": "file:../../creditcoin-js/creditcoin-js-v0.9.5.tgz",
+    "creditcoin-js": "file:../../creditcoin-js/creditcoin-js-v0.10.0.tgz",
     "prompts": "^2.4.2",
     "web3-validator": "^2.0.1"
   },
