+++ scripts/js/package.json
@@ -16,7 +16,7 @@
     "dependencies": {
         "@types/ws": "^8.5.3",
         "axios": "^1.1.2",
-        "creditcoin-js": "file:../../creditcoin-js/creditcoin-js-v0.9.5.tgz",
+        "creditcoin-js": "file:../../creditcoin-js/creditcoin-js-v0.10.0.tgz",
         "ws": "^8.5.0"
     },
     "devDependencies": {
